fix: circular dependency
- the refreshStatusBar in basicCommands.ts command imported inlineCompletionService.ts
which imported basicCommands.ts

- also the show quick pick command had circular dependency issues since it
  uses the codewhispererNode object. To fix this we moved the command to its
  own module
